Clean up FlintUISandbox a bit
